requirements.txt
requests==2.28.1
Flask==2.2.2
python-dotenv==0.21.0